requirements.txt
torch>=1.7
torchvision
numpy
matplotlib
Pillow
